+++ src/vue/forms/InvestForm.vue
@@ -342,7 +342,7 @@ export default {
 
       this.sale.quoteAssets.forEach(quote => {
         const balance = this.balances.find(balanceItem => {
-          return balanceItem.asset === quote.id
+          return balanceItem.asset === quote.asset.id
         })
 
         if (balance) {