+++ src/vue/modules/create-sale-form/mixins/load-assets.mixin.js
@@ -20,6 +20,11 @@ export default {
     baseAssets () {
       return this.assets.filter(a => a.isBaseAsset)
     },
+
+    defaultQuoteAsset () {
+      const asset = this.assets.find(a => a.isDefaultQuoteAsset)
+      return asset ? asset.code : ''
+    },
   },
 
   methods: {